Extract suggestions utility function and locate with Link UI
